--- src/main/java/chess/domain/piece/strategy/BishopMoveStrategy.java
@@ -1,16 +1,0 @@
-package chess.domain.piece.strategy;
-
-import chess.domain.position.Position;
-
-public class BishopMoveStrategy implements MoveStrategy {
-
-    @Override
-    public boolean isMovable(Position fromPosition, Position toPosition) {
-        return new DiagonalMoveStrategy().isMovable(fromPosition, toPosition);
-    }
-
-    @Override
-    public boolean isCatchable(Position fromPosition, Position toPosition) {
-        return isMovable(fromPosition, toPosition);
-    }
-}
