--- src/main/java/chess/domain/piece/strategy/MoveStrategy.java
@@ -1,9 +1,0 @@
-package chess.domain.piece.strategy;
-
-import chess.domain.position.Position;
-
-public interface MoveStrategy {
-    boolean isMovable(Position fromPosition, Position toPosition);
-
-    boolean isCatchable(Position fromPosition, Position toPosition);
-}
